// generated/src/app/api-models/pmb-ar-customer-master.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PmbArCustomerMaster {
  customerType: string;
  customerId: string;
  shortName: string;
  customerName1: string;
  customerName2: string;
  customerAddrLine1: string;
  customerAddrLine2: string;
  customerCity: string;
  customerState: string;
  customerPostalCode: string;
  useEftFlg: string;
  country: string;
  userDefined1: string;
  userDefined2: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}